% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{clear_geoscale_maps}
\alias{clear_geoscale_maps}
\title{Clear the registered spatial crosswalks}
\usage{
clear_geoscale_maps()
}
\value{
Invisibly \code{NULL}.
}
\description{
Mainly useful in tests.
}
\examples{
clear_geoscale_maps()
}
