% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{geo_get_rule}
\alias{geo_get_rule}
\title{Look up a registered rule}
\usage{
geo_get_rule(param)
}
\arguments{
\item{param}{Name of the value column.}
}
\value{
A list with elements \code{rule} and \code{weight}, or \code{NULL} if \code{param} has
not been registered.
}
\description{
Look up a registered rule
}
\examples{
geo_register_rule("demand", "sum")
geo_get_rule("demand")
geo_get_rule("not_registered")
}
