% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{geo_list_providers}
\alias{geo_list_providers}
\title{List registered providers}
\usage{
geo_list_providers()
}
\value{
A \code{data.frame} with columns \code{name} and \code{desc}.
}
\description{
List registered providers
}
\examples{
geo_list_providers()
}
