% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{geo_list_rules}
\alias{geo_list_rules}
\title{List registered rules}
\usage{
geo_list_rules()
}
\value{
A \code{data.frame} with columns \code{param}, \code{rule} and \code{weight}.
}
\description{
List registered rules
}
\examples{
geo_register_rule("invcost", "weighted_mean", weight = "km2")
geo_list_rules()
}
