% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoscale-class.R
\name{geo_weights}
\alias{geo_weights}
\title{Weight columns of a Geoscale}
\usage{
geo_weights(x)
}
\arguments{
\item{x}{A \code{\link{Geoscale}}.}
}
\value{
A character vector of weight column names (possibly empty).
}
\description{
Weight columns of a Geoscale
}
\examples{
geo_weights(geoscale_example())
}
