% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoscale-class.R
\name{geoscale_rank}
\alias{geoscale_rank}
\title{Level rank}
\usage{
geoscale_rank(x, level)
}
\arguments{
\item{x}{A \code{\link{Geoscale}}.}

\item{level}{Character vector of level names.}
}
\value{
An integer vector of ranks; \code{NA} for names that are not levels of
\code{x}.
}
\description{
Position of a level in the hierarchy: 1 is the coarsest.
}
\examples{
gs <- geoscale_example()
geoscale_rank(gs, c("zone", "atom"))
}
