% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoscales-package.R
\docType{package}
\name{geoscales-package}
\alias{geoscales}
\alias{geoscales-package}
\title{geoscales: Nested Regions and Spatial Hierarchies for Modeling}
\description{
Discrete spatial representations for optimization and simulation models,
including nested sets of regions, conversion between alternative spatial
resolutions, and mapping of data onto region hierarchies for modeling and
visualization.
}
\details{
Spatial companion to the
\href{https://github.com/optimal2050/timescales}{timescales} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://optimal2050.github.io/geoscales/r/}
  \item \url{https://optimal2050.github.io/geoscales/}
  \item \url{https://github.com/optimal2050/geoscales}
  \item Report bugs at \url{https://github.com/optimal2050/geoscales/issues}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

Authors:
\itemize{
  \item Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})
}

}
\keyword{internal}
