% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{get_geo_rule}
\alias{get_geo_rule}
\title{Look up a registered rule}
\usage{
get_geo_rule(param)
}
\arguments{
\item{param}{Name of the value column.}
}
\value{
A list with elements \code{rule} and \code{weight}, or \code{NULL} if \code{param} has
not been registered.
}
\description{
Look up a registered rule
}
\examples{
register_geo_rule("demand", "sum")
get_geo_rule("demand")
get_geo_rule("not_registered")
}
