% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{list_geo_providers}
\alias{list_geo_providers}
\title{List registered providers}
\usage{
list_geo_providers()
}
\value{
A \code{data.frame} with columns \code{name} and \code{desc}.
}
\description{
List registered providers
}
\examples{
list_geo_providers()
}
