% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{list_geoscale_rules}
\alias{list_geoscale_rules}
\title{List registered rules}
\usage{
list_geoscale_rules()
}
\value{
A \code{data.frame} with columns \code{param}, \code{rule} and \code{weight}.
}
\description{
List registered rules
}
\examples{
register_geoscale_rule("invcost", "weighted_mean", weight = "km2")
list_geoscale_rules()
}
