% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{geoscales-deprecated}
\alias{geoscales-deprecated}
\alias{geo_recast}
\alias{geo_filter}
\alias{geo_prune}
\alias{geo_attach_geometry}
\alias{geo_area}
\alias{geo_levels}
\alias{geo_rank}
\alias{geo_weights}
\alias{geo_regions}
\alias{geo_family}
\alias{geo_nests}
\alias{geo_ancestry}
\alias{geo_children}
\alias{geo_parents}
\alias{geo_descendants}
\alias{geo_ancestors}
\alias{geo_share}
\alias{geo_geometry}
\alias{geo_layout}
\alias{geo_autoplot}
\alias{geo_plot}
\alias{geo_register_rule}
\alias{geo_get_rule}
\alias{geo_list_rules}
\alias{geo_clear_rules}
\alias{geo_register_provider}
\alias{geo_provider}
\alias{geo_list_providers}
\title{Deprecated geoscales functions}
\usage{
geo_recast(...)

geo_filter(...)

geo_prune(...)

geo_attach_geometry(...)

geo_area(...)

geo_levels(...)

geo_rank(...)

geo_weights(...)

geo_regions(...)

geo_family(...)

geo_nests(...)

geo_ancestry(...)

geo_children(...)

geo_parents(...)

geo_descendants(...)

geo_ancestors(...)

geo_share(...)

geo_geometry(...)

geo_layout(...)

geo_autoplot(...)

geo_plot(...)

geo_register_rule(...)

geo_get_rule(...)

geo_list_rules(...)

geo_clear_rules(...)

geo_register_provider(...)

geo_provider(...)

geo_list_providers(...)
}
\arguments{
\item{...}{Arguments forwarded to the replacement function.}
}
\value{
See the replacement function.
}
\description{
These functions were renamed under the harmonized naming convention
shared with the timescales package. The old names warn and forward to
their replacements; they will be removed before the 1.0 release.
}
\details{
Data operations and object transforms (\code{verb_geoscale}):
\code{geo_recast()} -> \code{\link[=recast_geoscale]{recast_geoscale()}} (or the \code{recast()} generic),
\code{geo_filter()} -> \code{\link[=filter_geoscale]{filter_geoscale()}},
\code{geo_prune()} -> \code{\link[=prune_geoscale]{prune_geoscale()}},
\code{geo_attach_geometry()} -> \code{\link[=attach_geometry_geoscale]{attach_geometry_geoscale()}},
\code{geo_area()} -> \code{\link[=add_area_geoscale]{add_area_geoscale()}}.

Properties and queries (\verb{geoscale_*}):
\code{geo_levels()} -> \code{\link[=geoscale_levels]{geoscale_levels()}}, \code{geo_rank()} ->
\code{\link[=geoscale_rank]{geoscale_rank()}}, \code{geo_weights()} -> \code{\link[=geoscale_weights]{geoscale_weights()}},
\code{geo_regions()} -> \code{\link[=geoscale_regions]{geoscale_regions()}}, \code{geo_family()} ->
\code{\link[=geoscale_family]{geoscale_family()}}, \code{geo_nests()} -> \code{\link[=geoscale_nests]{geoscale_nests()}},
\code{geo_ancestry()} -> \code{\link[=geoscale_ancestry]{geoscale_ancestry()}}, \code{geo_children()} ->
\code{\link[=geoscale_children]{geoscale_children()}}, \code{geo_parents()} -> \code{\link[=geoscale_parents]{geoscale_parents()}},
\code{geo_descendants()} -> \code{\link[=geoscale_descendants]{geoscale_descendants()}}, \code{geo_ancestors()} ->
\code{\link[=geoscale_ancestors]{geoscale_ancestors()}}, \code{geo_share()} -> \code{\link[=geoscale_share]{geoscale_share()}},
\code{geo_geometry()} -> \code{\link[=geoscale_geometry]{geoscale_geometry()}}, \code{geo_layout()} ->
\code{\link[=geoscale_layout]{geoscale_layout()}}, \code{geo_autoplot()} -> \code{\link[=geoscale_autoplot]{geoscale_autoplot()}},
\code{geo_plot()} -> \code{\link[=geoscale_plot]{geoscale_plot()}}.

Registries:
\code{geo_register_rule()} -> \code{\link[=register_geo_rule]{register_geo_rule()}}, \code{geo_get_rule()} ->
\code{\link[=get_geo_rule]{get_geo_rule()}}, \code{geo_list_rules()} -> \code{\link[=list_geo_rules]{list_geo_rules()}},
\code{geo_clear_rules()} -> \code{\link[=clear_geo_rules]{clear_geo_rules()}},
\code{geo_register_provider()} -> \code{\link[=register_geo_provider]{register_geo_provider()}},
\code{geo_provider()} -> \code{\link[=get_geo_provider]{get_geo_provider()}}, \code{geo_list_providers()} ->
\code{\link[=list_geo_providers]{list_geo_providers()}}.
}
\keyword{internal}
